Moved 'old' Point upstream as it's replaced here by the new v2 Point.
