+++ source/imaer-shared/src/main/java/nl/overheid/aerius/geo/shared/EPSG.java
@@ -18,6 +18,8 @@ package nl.overheid.aerius.geo.shared;
 
 import java.io.Serializable;
 
+import nl.overheid.aerius.shared.domain.v2.geojson.Point;
+
 /**
  * EPSG map constants.
  */
