+++ source/imaer-shared/src/main/java/nl/overheid/aerius/shared/domain/v2/geojson/Point.java
@@ -22,7 +22,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 
 public class Point extends Geometry {
 
-  private static final long serialVersionUID = 2L;
+  private static final long serialVersionUID = 3L;
 
   private double[] coordinates = { 0, 0 };
 
@@ -49,6 +49,24 @@ public class Point extends Geometry {
     return coordinates[0];
   }
 
+  /**
+   * Get the rounded value of x.
+   * @return rounded x.
+   */
+  @JsonIgnore
+  public int getRoundedX() {
+    return (int) Math.round(getX());
+  }
+
+  /**
+   * Get the value of x rounded on centimeters.
+   * @return x rounded on centimeters
+   */
+  @JsonIgnore
+  public double getRoundedCmX() {
+    return ((int) Math.round(getX() * 100)) / 100.0;
+  }
+
   @JsonIgnore
   public void setX(final double x) {
     coordinates[0] = x;
@@ -57,6 +75,24 @@ public class Point extends Geometry {
   @JsonIgnore
   public double getY() {
     return coordinates[1];
+  }
+
+  /**
+   * Get the rounded value of y.
+   * @return rounded y.
+   */
+  @JsonIgnore
+  public int getRoundedY() {
+    return (int) Math.round(getY());
+  }
+
+  /**
+   * Get the value of y rounded on centimeters.
+   * @return y rounded on centimeters
+   */
+  @JsonIgnore
+  public double getRoundedCmY() {
+    return ((int) Math.round(getY() * 100)) / 100.0;
   }
 
   @JsonIgnore
