+++ source/imaer-shared/src/test/java/nl/overheid/aerius/geo/shared/BBoxTest.java
@@ -19,6 +19,8 @@ package nl.overheid.aerius.geo.shared;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
+import nl.overheid.aerius.shared.domain.v2.geojson.Point;
+
 /**
  * Test class for {@link BBox}.
  */
