+++ source/imaer-shared/src/test/java/nl/overheid/aerius/shared/domain/v2/geojson/PointTest.java
@@ -14,14 +14,14 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see http://www.gnu.org/licenses/.
  */
-package nl.overheid.aerius.geo.shared;
+package nl.overheid.aerius.shared.domain.v2.geojson;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.Test;
 
 /**
- * Test class for {@link Point}
+ * Unit test for {@link Point}.
  */
 public class PointTest {
